attempt 2 at conversion of bool to int

+++ src/compare_coef.py
@@ -38,8 +38,9 @@ def main(argv):
         res = pd.DataFrame({'Motif': motif, 
             'Var': andy_df.Cov, 
             'BetaDiff': andy_df.Estimate - jed_df.Estimate, 
-            'SignBeta': int(np.sign(andy_df.Estimate) == np.sign(jed_df.Estimate)), 
+            'SignBeta': np.sign(andy_df.Estimate) == np.sign(jed_df.Estimate), 
             'PDiff': andy_df.pval - jed_df.pval})
+        res.SignBeta = res.SignBeta.astype(int)
         with open(outfile, 'a') as f:
             res.to_csv(f, header = False, index = False)
     return(0)
